refactor(estimation): move R² arithmetic out of the result mixin
The accessor mixin keeps the thin `get_performance()` orchestration; the
goodness-of-fit arithmetic lives beside the other fit primitives in
`internals/performance_.py` so `Fegaussian` no longer imports a private
helper from a mixin module.
