+++ pyfixest/estimation/models/_result_accessor_mixin.py
@@ -1,6 +1,5 @@
 import functools
 import warnings
-from dataclasses import dataclass
 from importlib import import_module
 from typing import TYPE_CHECKING
 
@@ -19,83 +18,12 @@ from pyfixest.estimation.internals.literals import (
     InferenceType,
     _validate_literal_argument,
 )
+from pyfixest.estimation.internals.performance_ import (
+    PerformanceMeasures,
+    performance_measures,
+)
 from pyfixest.utils.dev_utils import _select_coefnames_and_indices
 from pyfixest.utils.utils import simultaneous_crit_val
-
-
-@dataclass(frozen=True, slots=True, kw_only=True)
-class PerformanceMeasures:
-    """Goodness-of-fit measures; the within variants are NaN without fixed effects."""
-
-    rmse: float
-    r2: float
-    adj_r2: float
-    r2_within: float
-    adj_r2_within: float
-
-
-def _performance_measures(
-    *,
-    Y: np.ndarray,
-    Y_within: np.ndarray,
-    residuals: np.ndarray,
-    weights: np.ndarray | None,
-    N: int | float,
-    k: int,
-    k_fe: int,
-    has_intercept: bool,
-    has_fixef: bool,
-) -> PerformanceMeasures:
-    """Compute R² measures from response-scale arrays and observation weights.
-
-    Parameters
-    ----------
-    Y : np.ndarray
-        Dependent variable, shape (N, 1).
-    Y_within : np.ndarray
-        Dependent variable residualized on the fixed effects, shape (N, 1),
-        in the units of ``Y``. Ignored when ``has_fixef`` is False.
-    residuals : np.ndarray
-        Residuals in the units of ``Y``, shape (N,).
-    weights : np.ndarray or None
-        User-scale observation weights, shape (N,) or (N, 1). ``None``
-        applies no weights.
-    N : int or float
-        Effective number of observations.
-    k : int
-        Number of estimated coefficients.
-    k_fe : int
-        Number of fixed-effect coefficients. Ignored when ``has_fixef`` is False.
-    has_intercept, has_fixef : bool
-        Whether the model has an intercept and fixed effects.
-    """
-    if weights is None:
-        ssu = np.sum(residuals**2)
-        ssy = np.sum((Y - np.mean(Y)) ** 2)
-    else:
-        w = weights.reshape((-1, 1))
-        ssu = np.sum(w.flatten() * residuals**2)
-        ssy = np.sum(w * (Y - np.average(Y, weights=w)) ** 2)
-
-    if has_fixef:
-        adj_factor = (N - has_intercept) / (N - k - k_fe)
-    else:
-        adj_factor = (N - has_intercept) / (N - k)
-
-    r2_within = adj_r2_within = np.nan
-    if has_fixef:
-        ssy_within = np.sum(Y_within**2) if weights is None else np.sum(w * Y_within**2)
-        adj_factor_within = (N - k_fe) / (N - k - k_fe)
-        r2_within = 1 - (ssu / ssy_within)
-        adj_r2_within = 1 - (ssu / ssy_within) * adj_factor_within
-
-    return PerformanceMeasures(
-        rmse=np.sqrt(ssu / N),
-        r2=1 - (ssu / ssy),
-        adj_r2=1 - (ssu / ssy) * adj_factor,
-        r2_within=r2_within,
-        adj_r2_within=adj_r2_within,
-    )
 
 
 class TidyColumnAccessors:
@@ -386,7 +314,7 @@ class ResultAccessorMixin(TidyColumnAccessors):
         fit._r2, fit._adj_r2, fit._r2_within
         ```
         """
-        measures = _performance_measures(
+        measures = performance_measures(
             Y=self._Y_untransformed.to_numpy(),
             Y_within=self._within_data.response,
             residuals=self._u_hat,
